+++ src/python/deepgnn/graph_engine/snark/tests/util_test.py
@@ -5,7 +5,7 @@
 from deepgnn.graph_engine.snark.decoders import JsonDecoder
 
 
-def json_to_linear_feature(features):
+def json_to_edge_list_feature(features):
     def get_f(f):
         if f is None:
             return "uint8 0"
@@ -27,7 +27,7 @@ def json_to_linear_feature(features):
     return " ".join(get_f(f) for f in features)
 
 
-def linear_encode(
+def edge_list_encode(
     node_id: int,
     blank: int,
     node_type: int,
@@ -37,7 +37,7 @@ def linear_encode(
     buffer: object = None,  # type: ignore
 ):
     """
-    Convert data to a line of the linear format.
+    Convert data to a line of the edge_list format.
 
     Parameters
     ----------
@@ -48,30 +48,30 @@ def linear_encode(
     edges: [(edge_src: int, edge_dst: int, edge_type: int, edge_weight: float, edge_features[ndarray, ...]), ...]
     """
     buffer.write(  # type: ignore
-        f"{node_id} -1 {node_type} {node_weight} {json_to_linear_feature(node_features)}\n"
+        f"{node_id} -1 {node_type} {node_weight} {json_to_edge_list_feature(node_features)}\n"
     )
     for edge_src, edge_dst, edge_type, edge_weight, edge_features in edges:
         buffer.write(  # type: ignore
-            f"{edge_src} {edge_dst} {edge_type} {edge_weight} {json_to_linear_feature(edge_features)}\n"
+            f"{edge_src} {edge_dst} {edge_type} {edge_weight} {json_to_edge_list_feature(edge_features)}\n"
         )
 
 
-def json_node_to_linear(node, buffer):
-    """Convert graph.json to graph.linear."""
+def json_node_to_edge_list(node, buffer):
+    """Convert graph.json to graph.csv."""
     gen = JsonDecoder().decode(node)
     node = next(gen)
     edges = [edge for edge in gen]
-    linear_encode(*node, edges, buffer)
+    edge_list_encode(*node, edges, buffer)
 
 
-def json_to_linear(filename_in, filename_out):
-    """Convert graph.json to graph.linear."""
+def json_to_edge_list(filename_in, filename_out):
+    """Convert graph.json to graph.csv."""
     file_in = open(filename_in, "r")
     file_out = open(filename_out, "w")
     while True:
         line = file_in.readline().strip()
         if not line:
             break
-        json_node_to_linear(line, file_out)
+        json_node_to_edge_list(line, file_out)
     file_in.close()
     file_out.close()